Update README.md
TMO, you are confusing resilience with availability in this context.

Tom Gilb: Resilience— the ability to absorb disturbances, to be changed and then to re-organise and still have the same identity (retain the same basic structure and ways of functioning).

So, apps keep on responding, even in the presence of failure. However, the *type* of response will reflect the general state of the system, e.g. success or failure mode. When the system is failing, it becomes unavailable—not being able to perform its normal or desired functioning—even while still responding to events (saying that it is unavailable). A resilient system will be able to quickly recover and jump back to desired behavior.

Therefore, availability and resilience imply four potential states of s system: available and resilient (preferred), unavailable and resilient (quick to snap back to available), available and delicate (intensive care required), unavailable and delicate (replace).

The receiving party (‘customer’ or ‘client’) will need to be prepared for these different types of responses or behaviors accordingly.