readable-reactivemanifesto
I reworded this slightly, changed minor grammatical errors (opinions), and made it so I could understand it. This is a great document, but is overloaded with unnecessary jargon making it too high-level for many. The Message-Driven section still needs a lot of work, and the line "commodity hardware and software platforms" as well as "composable properties" make this slightly more tech-focused than you want it to be.

I recognize most applications are in the software development world, but they don't have to be. And if this is to receive larger attention then these two issues should be addressed.