Update README.md
Responsive, resilient and elastic are brilliant design principles of which requirements can be derived. But "message-driven" is a solution to a part of these plus others. It just doesn't fit, it is a category mistake. It's a "how", not a "what". Remove it and it's wonderful. You still can advocate for message driven communication between loosly coupled micro systems or whatever, but is not on the same level as the other three. If we come up with a better way of achieving what we want to achieve with message driven architectures at the moment, like, say, neuronal one-way-high-speed connections or whatever, then responsive, resilient and elastic still make sense.