Update README.md
Resilience needs to be done through 'syncronization' not 'replication'; replication suggests duplicitous data. Within a database system, replication is equivalent to syncronization but the difference being that syncronization as a KEYWORD is significant in the fact that syncronized data is NOT replicated and/or duplicated; it is a representation of the same data across all systems.

If one must use the term 'replicated', one should use it in conjunction with 'syncronized' to be 100% clear.